[T] Changes in Task attributes: batch and max_instr
Tasks can now be batch and are "completed" since the start of the
execution. That means that if a normal task and a batch task are
running, the execution stops when the nomal task is complete, whitout
waiting for the batch task.

Also, max_megainstr is now max_instr and can't be defined in the "cmd"
dictionary.