[T] Max int instead of max time and PCM improvements
Instead of specifying the total execution time, now the user has to
specify the maximum number of intervals.

Also, the wrapper functions for PCM have been replaced by a class, which
receives funtions for resuming, pausing and waiting for tasks.

Finally, time is now repported in microseconds, and the PI controller
has been adjusted.