[F] Sfcoa: detect outlayers and min stall ratio th
If the most slowed down cluster has only one application and its the
distance of its stall cycles to average number of stall cycles is more
than 2 times de stddev, then use the maximum number of stall cycles from
the next cluster.

Also, do nothing if the ratio of stall cycles for the application with
more stall cycles is below a threshold.

Both options are configured using the configuration file.