[T] Replace all throw clauses with throw_with_trace
This functions throws the same exception but with an embedded stacktrace,
which can be retrieved later. Obviously, throws done inside external
libraries like STL do not embed this info, so calls to this libraries
that may throw and are considered of interest should be wrapped in a
try/catch to rethrow the exception with the stacktrace added.