[B] Improve children management (CAT and exiting)
Implemented a function to recursively found the descendants of a pid.

LinuxCAT did not take into account children processes when setting a
given pid into a CLOS. Now it does.

Also, under some circumstances the manager did not properly kill all the
processes it created. Now we have Herod the Great.